[svn] ConEmu.120930 sources and AutoUpdates information
git-svn-id: https://conemu-maximus5.googlecode.com/svn@1064 92116c04-ab5e-ca4c-3a8b-6f4eadf4b4a1

+++ src/ConEmuPlugin/ConEmuPlugin995.cpp
@@ -62,7 +62,7 @@ static wchar_t* GetPanelDir(HANDLE hPanel)
 	{
 		pszDir = (wchar_t*)calloc(nSize, sizeof(*pszDir));
 		if (pszDir)
-			nSize = InfoW995->Control(hPanel, FCTL_GETPANELDIR, nSize, (LONG_PTR)pszDir);
+			nSize = InfoW995->Control(hPanel, FCTL_GETPANELDIR, (DWORD)nSize, (LONG_PTR)pszDir);
 	}
 	_ASSERTE(nSize>0);
 	return pszDir;
